Got rid of history and extracted a helper
