+++ admin/src/helpers/routing/PrivateRoute.js
@@ -1,0 +1,28 @@
+import React from 'react';
+import { Route, Redirect } from 'react-router-dom';
+import session from '../Session';
+import PropTypes from 'prop-types';
+
+const PrivateRoute = ({ component: Component, ...rest }) => (
+  <Route
+    {...rest}
+    render={props =>
+      session.isLoggedIn() ? (
+        <Component {...props} />
+      ) : (
+        <Redirect
+          to={{
+            pathname: '/login',
+            state: { from: props.location }
+          }}
+        />
+      )
+    }
+  />
+);
+
+PrivateRoute.propTypes = {
+  component: PropTypes.any
+};
+
+export default PrivateRoute;
