+++ admin/src/routes.js
@@ -1,19 +1,14 @@
 import React from 'react';
-import { Component } from 'react';
-import PropTypes from 'prop-types';
 
-import { BrowserRouter as Router, Redirect, Route, Switch } from 'react-router-dom';
-
+import { BrowserRouter as Router, Route, Switch } from 'react-router-dom';
 // import App from './containers/App';
 import Login from './containers/Login/Login';
 import Dashboard from './components/core/pages/Dashboard';
 import NoMatch from './components/NoMatch';
-import history from './history';
-
-import session from './helpers/Session';
+import PrivateRoute from './helpers/routing/PrivateRoute';
 
 export default () => (
-  <Router history={history}>
+  <Router>
     <Switch>
       <Route path="/login" component={Login} exact />
       <PrivateRoute path="/dashboard" component={Dashboard} />
@@ -21,25 +16,3 @@ export default () => (
     </Switch>
   </Router>
 );
-
-const PrivateRoute = ({ component: Component, ...rest }) => (
-  <Route
-    {...rest}
-    render={props =>
-      session.isLoggedIn() ? (
-        <Component {...props} />
-      ) : (
-        <Redirect
-          to={{
-            pathname: '/login',
-            state: { from: props.location }
-          }}
-        />
-      )
-    }
-  />
-);
-
-PrivateRoute.propTypes = {
-  component: PropTypes.instanceOf(Component)
-};
